requirements.txt
Django==1.11.7
django-cors-headers==2.1.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
html5lib==1.0b10
mysqlclient==1.3.12
PyJWT==1.5.3
pytz==2017.3
six==1.11.0
webencodings==0.5.1
